Move event to shared and handle relay
