+++ Content.Shared/Temperature/TemperatureEvents.cs
@@ -1,0 +1,15 @@
+using Content.Shared.Inventory;
+
+namespace Content.Shared.Temperature;
+
+public sealed class ModifyChangedTemperatureEvent : EntityEventArgs, IInventoryRelayEvent
+{
+    public SlotFlags TargetSlots { get; } = ~SlotFlags.POCKET;
+
+    public float TemperatureDelta;
+
+    public ModifyChangedTemperatureEvent(float temperature)
+    {
+        TemperatureDelta = temperature;
+    }
+}
